livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[google,openai,silero,turn-detector]~=1.4",
    "livekit-plugins-cartesia",
    "livekit-plugins-deepgram",
    "livekit-plugins-openai",
    "livekit-plugins-silero",
    "livekit-plugins-google",
    "python-dotenv",
    "aiohttp",
    "google-genai",
    "livekit-plugins-noise-cancellation~=0.2",
]
